fixed with delete page and added a view page for posts using iframe also added cookie disclaimer

+++ Win21Pages/Pages/Posts/Delete.cshtml.cs
@@ -16,29 +16,34 @@ namespace Win21Pages.Pages.Posts
         private readonly ApplicationDbContext _userContext;
         private readonly UserManager<IdentityUser> _userManager;
         private readonly SignInManager<IdentityUser> _signInManager;
+        private readonly IWebHostEnvironment _env;
         public DeleteModel(ILogger<DeleteModel> logger,
             IDbContextFactory<PostContext> dbContextFactory,
             ApplicationDbContext userContext,
             UserManager<IdentityUser> userManager,
-            SignInManager<IdentityUser> signInManager)
+            SignInManager<IdentityUser> signInManager,
+            IWebHostEnvironment env)
         {
             _logger = logger;
             _dbContextFactory = dbContextFactory;
             _userContext = userContext;
             _userManager = userManager;
             _signInManager = signInManager;
+            _env = env;
         }
         [BindProperty(SupportsGet = true)]
         public IdentityUser SignedInUser { get; set; }
         [BindProperty(SupportsGet = true)]
         public Post? RemovablePost { get; set; }
-        
+        public string? PostUserId { get; set; }
         public string? Message { get; set; }
         public string? Test { get; set; }
-        public async Task OnGet(int postId)
+        public async Task OnGet(int postId, bool HasBeenDeleted)
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
                 SignedInUser = await _userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync();
+
+                
 
                 if (SignedInUser != null)
                 {
@@ -49,40 +54,54 @@ namespace Win21Pages.Pages.Posts
                             .Where(p => p.Id == postId)
                             .Where(u => u.UserId == SignedInUser.Id)
                             .FirstOrDefaultAsync();
+
+                        if (!HasBeenDeleted)
+                        {
+                        PostUserId = ctx.Posts.Find(postId).UserId;
+                        }
                     }
-
                 }
-
-                if (RemovablePost == null && SignedInUser == null)
+                if (RemovablePost == null && SignedInUser.Id != PostUserId)
                 {
                     Message = "ACCESS DENIED";
                 }
-
-                if (RemovablePost == null && SignedInUser != null)
+                if (HasBeenDeleted)
                 {
-                    //[TODO]: style Post Removed message
                     Message = "Post Removed!";
+                }                
+                if (HasBeenDeleted && RemovablePost != null)
+                {
+                    Message = "An unexpected error occurred... try again later...";
                 }
         }
-
-        public async Task<IActionResult> OnPostAsync(int id)  
+        public async Task<IActionResult> OnPostAsync(int id)
         {
-
-            //[TODO]: actually remove post from server somehow
+            RemoveDirectory removeDirectory = new RemoveDirectory();
+            bool deleted = false;
             if (id != null)
             {
+                var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                var currentUser = await _userManager.Users.Where(u => u.Id == userId).FirstOrDefaultAsync();
+
                 await using (var ctx = _dbContextFactory.CreateDbContext())
                 {
                     var DeletePost =  await ctx.Posts.FindAsync(id);
-                    if (DeletePost != null)
-                    {
-                        ctx.Posts.Remove(DeletePost);
-                        await ctx.SaveChangesAsync();
-                    }
 
+                    if (DeletePost != null && DeletePost.UserId == currentUser.Id)
+                    {
+                        var deletePath = Path.Combine(
+                            _env.ContentRootPath,"UserData", currentUser.Id, DeletePost.PostName);
+                        var IsRemoved = await removeDirectory.DeleteDirectory(deletePath);
+                        if (IsRemoved)
+                        {
+                            ctx.Posts.Remove(DeletePost);
+                            await ctx.SaveChangesAsync();
+                            deleted = IsRemoved;
+                        }
+                    }
                 }
             }
-            return RedirectToPage();
+            return RedirectToRoute("", new {postId = id, HasBeenDeleted = deleted });
         }
     }
 }
